# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "商品微服务相关接口",
        "title": "商品服务 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/product-ms/v1",
    "paths": {
        "/customer/cart": {
            "get": {
                "description": "Get user's cart info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get user's cart info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.CartListVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/items": {
            "post": {
                "description": "Create a cart item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Create a cart item",
                "parameters": [
                    {
                        "description": "cart item info",
                        "name": "cart_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CartItemBasicVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.CartItemBasicVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/items/:item_id": {
            "put": {
                "description": "Update a cart item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Update a cart item",
                "parameters": [
                    {
                        "description": "cart item info",
                        "name": "cart_item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CartItemBasicVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.CartItemBasicVO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a cart item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Delete a cart item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cart item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/price-estimate": {
            "get": {
                "description": "Calculate order price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Calculate order price",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.CartPriceEstimateResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/selected-num": {
            "get": {
                "description": "Get number of selected items in cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart"
                ],
                "summary": "Get number of selected items in cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/data.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "err_msg": {
                    "type": "string"
                }
            }
        },
        "data.CartItemBasicVO": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                },
                "selected": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "data.CartItemDetailVO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "product_info": {
                    "$ref": "#/definitions/types.ProductInfo"
                },
                "quantity": {
                    "type": "integer"
                },
                "selected": {
                    "type": "boolean"
                },
                "status": {
                    "description": "1: normal, 2: out of stock",
                    "type": "integer"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
        "data.CartListVO": {
            "type": "object",
            "properties": {
                "cart_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.CartItemDetailVO"
                    }
                },
                "selected_item_count": {
                    "type": "integer"
                },
                "selected_price": {
                    "type": "integer"
                }
            }
        },
        "data.CartPriceEstimateResult": {
            "type": "object",
            "properties": {
                "product_price": {
                    "type": "integer"
                },
                "shipping_price": {
                    "type": "integer"
                },
                "tax": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "types.ProductInfo": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "string"
                },
                "care_instructions": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "string"
                },
                "material": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pic_info": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "status": {
                    "description": "0: 未上架, 1: 已上架",
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "weight": {
                    "type": "string"
                }
            }
        },
...
    }
}